Update perf-tests x2
